+++ CHANGELOG.md
@@ -24,11 +24,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `Removed` for now removed features.
 * `Fixed` for any bug fixes.
 * `Security` in case of vulnerabilities.
- -->
+-->
 
-## [0.0.1] - 2020-09-02
+## [v0.1.0] - 2020-09-02
+
+### Add
+
+* `NonEmptyVec`
+
+### Breaking changes
+
+* `as_mut_ptr` and `as_mut_slice` now require `&mut self`.
+* Remove `Clone` impl for `NonEmptyMutSlice`.
+  It is kinda unsound for multiple mutable references.
+
+## [v0.0.1] - 2020-09-02
 
 First release
 
-<!-- [0.1.3]: https://github.com/lzutao/rust-oom/compare/v0.0.1...v0.1.3 -->
-[0.0.1]: https://github.com/lzutao/rust-oom/releases/tag/v0.0.1
+[v0.1.0]: https://github.com/lzutao/rust-oom/compare/v0.0.1...v0.1.0
+[v0.0.1]: https://github.com/lzutao/rust-oom/releases/tag/v0.0.1